Move PowerSupply from __init__.py to power.py
