+++ ev3dev/__init__.py
@@ -309,93 +309,3 @@ def list_devices(class_name, name_pattern, **kwargs):
 
     return (Device(class_name, name, name_exact=True)
             for name in list_device_names(classpath, name_pattern, **kwargs))
-
-
-class PowerSupply(Device):
-
-    """
-    A generic interface to read data from the system's power_supply class.
-    Uses the built-in legoev3-battery if none is specified.
-    """
-
-    SYSTEM_CLASS_NAME = 'power_supply'
-    SYSTEM_DEVICE_NAME_CONVENTION = '*'
-    __slots__ = [
-    '_measured_current',
-    '_measured_voltage',
-    '_max_voltage',
-    '_min_voltage',
-    '_technology',
-    '_type',
-    ]
-
-    def __init__(self, address=None, name_pattern=SYSTEM_DEVICE_NAME_CONVENTION, name_exact=False, **kwargs):
-
-        if address is not None:
-            kwargs['address'] = address
-        super(PowerSupply, self).__init__(self.SYSTEM_CLASS_NAME, name_pattern, name_exact, **kwargs)
-
-        self._measured_current = None
-        self._measured_voltage = None
-        self._max_voltage = None
-        self._min_voltage = None
-        self._technology = None
-        self._type = None
-
-    @property
-    def measured_current(self):
-        """
-        The measured current that the battery is supplying (in microamps)
-        """
-        self._measured_current, value = self.get_attr_int(self._measured_current, 'current_now')
-        return value
-
-    @property
-    def measured_voltage(self):
-        """
-        The measured voltage that the battery is supplying (in microvolts)
-        """
-        self._measured_voltage, value = self.get_attr_int(self._measured_voltage, 'voltage_now')
-        return value
-
-    @property
-    def max_voltage(self):
-        """
-        """
-        self._max_voltage, value = self.get_attr_int(self._max_voltage, 'voltage_max_design')
-        return value
-
-    @property
-    def min_voltage(self):
-        """
-        """
-        self._min_voltage, value = self.get_attr_int(self._min_voltage, 'voltage_min_design')
-        return value
-
-    @property
-    def technology(self):
-        """
-        """
-        self._technology, value = self.get_attr_string(self._technology, 'technology')
-        return value
-
-    @property
-    def type(self):
-        """
-        """
-        self._type, value = self.get_attr_string(self._type, 'type')
-        return value
-
-    @property
-    def measured_amps(self):
-        """
-        The measured current that the battery is supplying (in amps)
-        """
-        return self.measured_current / 1e6
-
-    @property
-    def measured_volts(self):
-        """
-        The measured voltage that the battery is supplying (in volts)
-        """
-        return self.measured_voltage / 1e6
